VSTS build status with a GitHub README file
Create a VSTS build status with a GitHub README file
This section explores possibilities for further integrating VSTS and GitHub. You will create a notification for the VSTS build definition by populating a GitHub readme file with Markdown that points to the build notification URL. This same Markdown can link back to the build summary page in VSTS.

Navigate to your GitHub account. Select Code. Create a README.md file unless one already exists.
For this step, use the URL from step 13 in the previous section. Modify the README.md file by adding Markdown and an img tag, as shown below. Create a Markdown link that displays the build status image and links to the build summary in VSTS. Modify the below example with your account and build ID information.

[![Build Status](https://{your-account}.visualstudio.com/_apis/public/build/definitions/{guid}/{id}/badge)](https://{your-account}.visualstudio.com/{your-project}/_build/index?definitionId={id})